requirements.txt
python-dotenv==0.21.0
Authlib~=1.0.0
Django==4.1
django-filter==22.1
djangorestframework==3.13.1
django-cors-headers==3.13.0
requests==2.22.0
django-storages==1.13.1
boto3==1.24.84
Markdown==3.4.1
psycopg2==2.9.3
drf-yasg==1.21.4
gunicorn==20.1.0
